Sumamos iconos de redes sociales,cambiamos el detalles de los derechos reservados,añadimos una linea separadora
