+++ ProyectoFinalRC/package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.10.2",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.52.0",
+    "react-icons": "^5.2.1",
     "react-multi-carousel": "^2.8.5",
     "react-responsive": "^10.0.0",
     "react-router-dom": "^6.23.1"
