+++ ProyectoFinalRC/src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
 import HomePage from './Components/Pages/Home/HomePage/HomePage';
 import Navbar from './Components/Layout/Header/NavBar/NavBar';
-import Footer from './Components/Layout/Footer/Footer/Footer';
+import Footer from './Components/Layout/Footer/Footer';
 
 function App() {
   return (
