+++ ProyectoFinalRC/src/Components/Layout/Footer/Footer.css
@@ -1,8 +1,9 @@
 /* src/Components/Footer/Footer.css */
 .footer {
     background-color: #202020;
-    color: white;
+    color: rgb(185, 180, 180);
     padding: 40px 8% 0px 8%;
+    
 }
 .footer .container {
     padding: 0 8%; /* Estilo responsive para padding lateral */
@@ -16,22 +17,41 @@
     margin-bottom: 10px;
 } 
 .footer p, .footer ul, .footer li, .footer a {
-    color: #ffffff;
+    color: rgb(185, 180, 180);
     text-decoration: none;
     margin: 0;
     padding: 0;
     font-weight: 500;
 }  
-.footer a:hover {
-    text-decoration: underline;
-}  
+
 .footer .list-unstyled {
     padding-left: 0;
-    list-style: none;
+    list-style: none;   
 }
 .columContainer{
     margin-bottom: 40px;
 }  
+.link-footer{
+    text-decoration: none;
+
+}
+.link-footer:hover{
+    text-decoration: none;
+    color: rgb(255, 255, 255);
+}
+.icon-sociales{
+    color:white;
+    background-color: orangered;
+    padding: 5px;
+    border-radius: 25%;
+}
+.icon-sociales:hover{
+    background-color: rgb(199, 92, 54);
+    color: rgb(211, 205, 205);
+}
+.titulo-lista-footer{
+    color: rgb(230, 226, 226);
+}
 @media (max-width: 768px) {
     .footer .container {
       padding: 0 5%;
